fix : 동요 생성 수정

See merge request s12-final/S12P31B104!161

+++ backend/aieng/src/main/java/com/ssafy/aieng/domain/theme/dto/response/ThemeResponse.java
@@ -10,7 +10,7 @@ public class ThemeResponse {
     private String themeKo;
     private String themeEn;
     private String imageUrl;
-    private Byte totalWords;
+    private Integer totalWords;
 
     public static ThemeResponse from(Theme theme) {
         return ThemeResponse.builder()
